Add new composer: adder
That combine two or more controllers, such that the output of the
composer is the sum of the outputs for each controller.

With the adder and the basic controllers (proportional, derivative, and
integral), it's possible to create complex controller, for instance:

* PD (proportional + derivative)
* PI (proportional + integral)
* PID (proportional + derivative + integral)

To simplify the iteration over the std::tuple of controllers, the
tuple_helper was created.